Move to its own Context and set methods on Drupal cores
